POST To enable the endpoint to have the ability for users to add a new item, we must implement the POST action using the [HttpPost]. By passing the item (in this example, a Pizza) into the method as a parameter, ASP.NET Core will automatically convert any application/json that is sent to the endpoint into a populated .NET Pizza object.
Here's the method signature of the Create method you'll be implementing in the next section:
The [HttpPost] attribute will map HTTP POST requests sent to http://localhost:5000/pizza the Create() method. Instead of returning a list of pizzas, as we saw with the Get() method, this method will return an IActionResult response, which will let the client know if the request succeeded and the ID of the newly created pizza. This is done using standard HTTP status codes, so it will easily integrate with clients regardless of the language or platform they're running on.

POST
ASP.NET Core
action result HTTP status code Description
CreatedAtAction 201 The pizza was added to the in-memory cache.
The pizza is included in the response body in the media type as defined in the accept HTTP request header (JSON by default).
BadRequest is implied 400 The request body's pizza object is invalid.
Fortunately, the ControllerBase has utility methods that will create the appropriate HTTP response codes and messages for us. You'll see how those work in the next exercise.