The preceding action:
Responds only to the HTTP POST verb, as denoted by the [HttpPost] attribute.
Inserts the request body's Pizza object into the in-memory cache.
 Note

Because the controller is annotated with the [ApiController] attribute, it's implied that the Pizza parameter will be found in the request body.

The first parameter in the CreatedAtAction method call represents an action name. The nameof keyword is used to avoid hard-coding the action name. CreatedAtAction uses the action name to generate a location HTTP response header with a URL to the newly created pizza, as explained in the previous unit.