Remove a pizza Let's enable a pizza to be removed through our web API with a DELETE method.
Replace the // DELETE action comment in Controllers/PizzaController.cs with the following code:

The preceding action:

Responds only to the HTTP DELETE verb, as denoted by the [HttpDelete] attribute.
Requires that id parameter's value is included in the URL segment after pizza/.
Returns IActionResult because the ActionResult return type isn't known until runtime. The NotFound and NoContent methods return NotFoundResult and NoContentResult types, respectively.
Queries the in-memory cache for a pizza matching the provided id parameter.